requirements.txt
streamlit==1.30.0
streamlit-drawable-canvas==0.9.3
pillow
numpy
